@@ -163,12 +163,14 @@ class _BarcodeReaderPageState extends State<BarcodeReaderPage> {
   }
 
   Widget _buildCancelButton() {
-    return FlatButton(
+    return TextButton(
       onPressed: () {
         _captureController.pause();
         Navigator.of(context).pop();
       },
-      textColor: widget.buttonColor,
+      style: TextButton.styleFrom(
+        foregroundColor: widget.buttonColor,
+      ),
       child: Text(widget.cancelButtonText),
     );
   }
